requirements.txt
aiohttp==3.6.2
aiohttp-sse-client==0.1.6
alembic==1.3.0
asn1crypto==1.3.0
async-timeout==3.0.1
attrs==19.3.0
boto3==1.11.17
botocore==1.14.17
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
coincurve==13.0.0
crc16==0.1.1
cytoolz==0.10.1
DateTime==4.3
dnspython==1.16.0
docutils==0.15.2
eciespy==0.3.4
ed25519==1.5
electrumsv-secp256k1==0.9.1
eth-hash==0.2.0
eth-keys==0.3.1
eth-typing==2.2.1
eth-utils==1.8.4
eventlet==0.25.2
falcon==2.0.0
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.11.0
google-auth-httplib2==0.0.3
greenlet==0.4.15
gunicorn==20.0.2
httplib2==0.17.0
idna==2.8
idna-ssl==1.1.0
iniherit==0.3.9
jmespath==0.9.4
Mako==1.1.1
MarkupSafe==1.1.1
mnemonic==0.19
monotonic==1.5
multidict==4.7.4
numpy==1.18.1
oauthlib==3.1.0
pbkdf2==1.3
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pycryptodome==3.9.6
PyJWT==1.7.1
PyNaCl==1.3.0
python-dateutil==2.8.1
python-dotenv==0.11.0
python-editor==1.0.4
python-twitter==3.5
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
six==1.14.0
SQLAlchemy==1.3.13
sqlparse==0.3.0
sseclient==0.0.24



text-unidecode==1.3
toml==0.10.0
toolz==0.10.0
typing-extensions==3.7.4.1
uritemplate==3.0.1
urllib3==1.25.8
voluptuous==0.11.7
yarl==1.4.2
yoyo-migrations==7.0.1
zope.interface==4.7.1
